backend/requirements.txt
fastapi
uvicorn
pydantic
openai
pytest
httpx
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
